backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.0
python-dotenv==1.0.1
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.8
pypdf==4.2.0
sentence-transformers==3.0.0
chromadb==0.5.3
openai==1.35.1